feature(frontend): Extract publish to hub button
